Add a module folder and importing separated files to index.js
