+++ src/modules/EditData.js
@@ -1,0 +1,8 @@
+import storeLocalStorage from './storeLocalStorage.js';
+
+const edit = (storeTasks, description, index) => {
+  storeTasks[index].description = description;
+  storeLocalStorage(storeTasks);
+};
+
+export default edit;
