+++ src/modules/clearAll.js
@@ -1,0 +1,6 @@
+const clearAll = (storeTasks) => {
+  storeTasks = storeTasks.filter((element) => !element.completed);
+  return storeTasks;
+};
+
+export default clearAll;
