+++ src/modules/storeLocalStorage.js
@@ -1,0 +1,5 @@
+const storeLocalStorage = (storeTasks) => {
+  localStorage.setItem('Tasks', JSON.stringify(storeTasks));
+};
+
+export default storeLocalStorage;
